backend/pyproject.toml
[project]
name = "lazy-mcp"
version = "0.1.0"
description = "Backend for Azure LLM and MCP integration"
authors = ["Max Ostrowski <max@example.com>"]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
    "pydantic",
    "anyio"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
